requirements.txt
# Core ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
matplotlib==3.9.0
seaborn==0.13.2
boto3==1.34.87
prometheus-fastapi-instrumentator==6.1.0
pytest==8.2.0
httpx==0.27.0
python-dotenv==1.0.1
